+++ CHANGELOG.md
@@ -12,6 +12,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `RECEIPT_ID` option to external id generation for forms
 - `postSubmissionReceipt` property to form validation
 - `hintPosition` property to element validation
+- `toEmail`, `ccEmail` and `bccEmail` to `emailSubmissionEventConfiguration`
+
+### Changed
+
+- Deprecated `emailSubmissionEventConfiguration.email`
 
 ## [2.0.2] - 2023-03-13